BF - fix broadcasting of area weights
The area_weights function no longer raises ValueError referring to
broadcasting of shapes when the lat/lon grid does not occupy the
rightmost dimensions of the cube.

Support for lat/lon dimensions that are not next to one another
is included.

Support for weights generated from scalar coordinates is included.