BUG: Don't check coding standard for build dir
Currently pep8 checks the build directory for changes contrary the
ignore the list.  Directory specific checks now implemented.  This is
to overcome the sensitivity of the pep8 compliance check
(test_coding_standards) on whether the
documentation has been built before or after the test (i.e. a failure
caused by the built documentation).