BUG FIX: Serialize collapsed coordinates by datatype, not units
This fixes a bug introduced by #847, which is encountered when attempting to
collapse along a dimension coordinate with unknown units. In such a case, the
coordinate value is converted into a string, which fails for dimension
coordinates because dimension coordinates must be numeric.

In contrast, this new patch chooses how to serialize a collapsed coordinate by
its datatype instead of its units. String point values are joined by '|',
whereas all other point values (which are presumed to be numeric) are replaced
by the mean.