doc: Cleanup and fix front page's random image view.
Using the already-generated random_image.js, shuffle list of images, and
then pick the first 5 for display. Since the image paths are no longer
hard-coded, this should work with matplotlib 1.2 and 1.3+.

Additionally, fix layout of the sample images. The links and the images
are 'float'd but there is no 'clear'ing of the floating, so the body
actually ends before them.

Finally, CSS-ify bullets under the slideshow.