BUG: Lazy cube with mask not saved correctly to NetCDF
Currently the fill_value keyword argument to CFVariable is not utilised
when writting a NetCDF file where the data is lazy.  There is the issue
of how we determine whether to pass a fill_value of None since we cannot
check if its a masked array or not without touching the data and biggus
returns a default fill_value either way.