Add my fix to use cell areas and add representation
Still needs to be cleaned

Also not sure what is meant to happen with the cell measure after collapsing.
It is causing issues for representation (scalar cell measure not yet handled)
and it is not clear what the behaviour should be. Should the cell measure
simply disappear, should we keep the average cell measure, the total cell
measure?