Provide a unittest for properly deleting a CFReader during failure
This test is slightly complicated: it tests that no errors occur
during garbage collection of the CFReader; that is, during use of the
`__del__` method.

Any errors that may occur, however, will not be raised, but just
printed to standard error, as a safety measure (since a destructor
should always run and complete). Thus, the standard error is captured
and its length verified to be zero.

For the actual test, we create a netCDF4 file with a proper signature
(HDF5 actually), but otherwise broken; the correct signature (magic
number) is needed for use in the indirect `load_cubes` call, which
first attempts to determine the file type before instantiating a
`CFReader` object.

A final remark: this test does not fit very well in the unittest
class, since it doesn't rely on external data. For clarity, however,
it is kept within the TestCFReader class.